+++ lib/puppet-check.rb
@@ -14,6 +14,9 @@ class PuppetCheck
   # initialize output format option
   @output_format = 'text'
 
+  # initialize octocatalog-diff options
+  @octoconfig = '.octocatalog-diff.cfg.rb'
+
   # initialize diagnostic output arrays
   @error_files = []
   @warning_files = []
@@ -26,7 +29,7 @@ class PuppetCheck
 
   # allow the parser methods read user options and append to the file arrays; allow CLI and tasks write to user options
   class << self
-    attr_accessor :future_parser, :style_check, :regression_check, :output_format, :error_files, :warning_files, :clean_files, :ignored_files, :puppetlint_args, :rubocop_args
+    attr_accessor :future_parser, :style_check, :regression_check, :output_format, :octoconfig, :error_files, :warning_files, :clean_files, :ignored_files, :puppetlint_args, :rubocop_args
   end
 
   # main runner for PuppetCheck